+++ config/routes.rb
@@ -345,7 +345,6 @@ Rails.application.routes.draw do
 
       resources :design_options do
         collection do
-          get :new_section
           get :new_variable
           get :new_existing_variable
           post :create_section