+++ db/migrate/20190326140806_remove_image_from_sections.rb
@@ -1,0 +1,5 @@
+class RemoveImageFromSections < ActiveRecord::Migration[6.0]
+  def change
+    remove_column :sections, :image, :string
+  end
+end